+++ console/src/components/Services/Data/RawSQL/RawSQL.js
@@ -33,7 +33,6 @@ import { getSourceDriver } from '../utils';
 import { getDataSources } from '../../../../metadata/selector';
 import { services } from '../../../../dataSources/services';
 import { isFeatureSupported } from '../../../../dataSources';
-
 /**
  * # RawSQL React FC
  * ## renders raw SQL page on route `/data/sql`
@@ -86,7 +85,8 @@ const RawSQL = ({
   useEffect(() => {
     const driver = getSourceDriver(sources, selectedDatabase);
     setSelectedDriver(driver);
-    if (driver !== 'postgres') setStatementTimeout(null);
+    if (isFeatureSupported('rawSQL.statementTimeout'))
+      setStatementTimeout(null);
   }, [selectedDatabase, sources]);
 
   const dropDownSelectorValueChange = value => {
@@ -483,7 +483,7 @@ const RawSQL = ({
           {getMetadataCascadeSection()}
           {getMigrationSection()}
 
-          {selectedDriver === 'postgres' && (
+          {isFeatureSupported('rawSQL.statementTimeout') && (
             <StatementTimeout
               statementTimeout={statementTimeout}
               isMigrationChecked={