+++ console/src/dataSources/services/mysql/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DataSourcesAPI } from '../..';
-import { Table, TableColumn } from '../../types';
+import { Table, TableColumn, ViolationActions } from '../../types';
 import {
   getAlterForeignKeySql,
   getCreateFKeySql,
@@ -131,6 +131,14 @@ const operators = [
 
 const createSQLRegex = /create\s*((?:|or\s*replace)\s*view|\s*(table|function|view))\s*(?:\s*if*\s*not\s*exists\s*)?(((\`?\w+\`?)\.(\`?\w+\`?))|(\`?\w+\`?))/g; // eslint-disable-line
 
+const violationActions: ViolationActions[] = [
+  'restrict',
+  'no action',
+  'cascade',
+  'set null',
+  'set default',
+];
+
 export const mysql: DataSourcesAPI = {
   getFunctionSchema: () => {
     return '';
@@ -236,4 +244,5 @@ WHERE
   viewsSupported: false,
   supportedColumnOperators: null,
   aggregationPermissionsAllowed: false,
+  violationActions,
 };