+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## Next release
 
-
 ### Bug fixes and improvements
 (Add entries below in the order of server, console, cli, docs, others)
 
 - console: add `pool_timeout`, `connection_lifetime` and `isolation_level` connection params to connect database form
 - console: add check constraints and comments to MS SQL Server tables' read-only modify page
+- console: add create table functionality for MS SQL Server 
 
 ## v2.0.0-alpha.11